+++ docs/source/history.rst
@@ -24,6 +24,8 @@ Removed *self* argument from *autodoc* methods.
 
 Add a new function to save current state: :func:`save_config`.
 
+No sleep in :func:`I2C_write` loop.
+
 
 V1.5
 --------